ploop_set_max_delta_size(): fix definition
On POWER, __u64 is defined as "unsigned long", while on x86_64
it is "unsigned long long", thus the following error:

> ploop.c:3431:5: error: conflicting types for ‘ploop_set_max_delta_size’
>  int ploop_set_max_delta_size(struct ploop_disk_images_data *di, __u64 size)
>      ^
> In file included from ploop.h:13:0,
>                  from ploop.c:39:
> ../include/libploop.h:260:5: note: previous declaration of ‘ploop_set_max_delta_size’ was here
>  int ploop_set_max_delta_size(struct ploop_disk_images_data *di, unsigned long long size);
>      ^

Fix function definition to match its prototype.

+++ lib/ploop.c
@@ -3392,7 +3392,7 @@ int ploop_get_spec(struct ploop_disk_images_data *di, struct ploop_spec *spec)
 	return ret;
 }
 
-int ploop_set_max_delta_size(struct ploop_disk_images_data *di, __u64 size)
+int ploop_set_max_delta_size(struct ploop_disk_images_data *di, unsigned long long size)
 {
 	char conf[PATH_MAX];
 	char dev[64];
